+++ src/settings.ts
@@ -1,6 +1,5 @@
-import {moduleName, TIME_ZONES} from "./const.js";
+import {moduleName} from "./const.js";
 import {localize} from "./utils.js";
-import {rerenderTime} from "./index.js";
 
 abstract class SubSettings extends FormApplication {
 
@@ -92,49 +91,6 @@ class AutomationSettings extends SubSettings {
     }
 }
 
-
-class CalendarSettings extends SubSettings {
-
-    static _namespace = "calendar";
-
-    static get settings() {
-        return {
-            showCalendar: {
-                name: `${moduleName}.SETTINGS.showCalendar.name`,
-                hint: `${moduleName}.SETTINGS.showCalendar.hint`,
-                default: false,
-                type: Boolean,
-                onChange: (_value: any) => {
-                    rerenderTime(ui.players.element)
-                },
-            },
-            calendarFormatStyle: {
-                name: `${moduleName}.SETTINGS.calendarFormatStyle.name`,
-                hint: `${moduleName}.SETTINGS.calendarFormatStyle.hint`,
-                type: String,
-                choices: {
-                    "en-US": game.i18n.localize(`${moduleName}.SETTINGS.calendarFormatStyle.us`),
-                    "en-GB": game.i18n.localize(`${moduleName}.SETTINGS.calendarFormatStyle.eu`),
-                },
-                default: "en-GB",
-                onChange: (_value: any) => {
-                    rerenderTime(ui.players.element)
-                },
-            },
-            currentTimeZone: {
-                name: `${moduleName}.SETTINGS.currentTimeZone.name`,
-                hint: `${moduleName}.SETTINGS.currentTimeZone.hint`,
-                type: String,
-                default: 'utc',
-                choices: TIME_ZONES,
-                onChange: (_value: any) => {
-                    rerenderTime(ui.players.element)
-                },
-            }
-        };
-    }
-}
-
 export class Settings {
 
     static get(name: string) {
@@ -152,15 +108,5 @@ export class Settings {
             restricted: true
         });
         AutomationSettings.init();
-
-        game.settings.registerMenu(moduleName, "calendar", {
-            name: localize(`${moduleName}.SETTINGS.Menu.calendar.name`),
-            label: localize(`${moduleName}.SETTINGS.Menu.calendar.label`),
-            hint: "",
-            icon: "fa-solid fa-clock",
-            type: CalendarSettings,
-            restricted: true
-        });
-        CalendarSettings.init();
     }
 }